Refactors existing test, removes unecessary comments and lines
